Add writing of incomplete beginning of a fox 🦊
Previously, running a command such as
```
dd if=/dev/foxes bs=3 count=12
```
would output zero foxes. This was not the intended behaviour, as the
total requested amount of bytes (`3*12=36=9*4`) was enough to accomodate
for 9 foxes. This is now fixed with this patch that considers the
remaining bytes available to write in the `IoBufferWriter`, and outputs
however many bytes of the beginning of the next fox remain.